Refactor Forgot page to conditionally render based on user session
Modified the Forgot page to dynamically render either the New Password page or the Forgot Password page depending on the user's session status.
If the user has an active session, the New Password page is displayed.
If the user does not have an active session, the Forgot Password page is displayed.
This change improves the user experience by ensuring that users are directed to the appropriate page based on their login status.